+++ test/integration/app.test.js
@@ -1,5 +1,4 @@
-import ldap from 'ldapjs';
-jest.mock('ldapjs');
+import {Client as Connection} from 'ldapts';
 import request from 'supertest';
 import jwt from 'jsonwebtoken';
 import {config} from '../../src/config';
@@ -12,14 +11,18 @@ const fixtures = {
       'cn=test,dc=example,dc=com',
     ],
   },
-  tokenPayload: {
+  tokenReviewPayload: {
+    username: 'john.doe',
+    iat: Math.floor(Date.now() / 1000),
+    exp: Math.floor(Date.now() / 1000 + 3600),
+  },
+  tokenReviewResponse: {
     username: 'john.doe',
     uid: 'john.doe',
     groups: [
       'test',
     ],
-    iat: Math.floor(Date.now() / 1000),
-    exp: Math.floor(Date.now() / 1000 + 3600),
+    extra: {},
   },
   tokenReviewTemplate: {
     apiVersion: 'authentication.k8s.io/v1beta1',
@@ -30,16 +33,12 @@ const fixtures = {
   },
 };
 
-let connection = ldap.createClient();
+let connection = new Connection();
 
 describe('test app routes', () => {
   beforeEach(() => {
-    connection.starttlsReturnsError = false;
     connection.bindReturnsError = false;
     connection.searchReturnsError = false;
-    connection.searchEmitsError = false;
-    connection.searchEmitsEnd = false;
-    connection.searchEmitsEndStatus = 0;
     connection.searchResult = {
       uid: fixtures.ldap.username,
       memberOf: fixtures.ldap.memberOf,
@@ -69,9 +68,7 @@ describe('test app routes', () => {
       })
       .expect((response) => {
         let object = jwt.decode(response.text);
-        expect(object.username).toBe(fixtures.tokenPayload.username);
-        expect(object.uid).toBe(fixtures.tokenPayload.uid);
-        expect(object.groups).toEqual(fixtures.tokenPayload.groups);
+        expect(object.username).toBe(fixtures.tokenReviewPayload.username);
       })
       .end((error) => {
         done(error);
@@ -104,19 +101,9 @@ describe('test app routes', () => {
         done(error);
       });
   });
-  test('GET /auth: 500 Internal Server Error', (done) => {
-    connection.searchResult.memberOf = null;
-    request(app)
-      .get('/auth')
-      .auth('john.doe', 'secret')
-      .expect(500)
-      .end((error) => {
-        done(error);
-      });
-  });
   test('POST /token: 200 OK', (done) => {
     let tokenReview = fixtures.tokenReviewTemplate;
-    tokenReview.spec.token = jwt.sign(fixtures.tokenPayload, config.jwt.key);
+    tokenReview.spec.token = jwt.sign(fixtures.tokenReviewPayload, config.jwt.key);
     request(app)
       .post('/token')
       .send(tokenReview)
@@ -126,7 +113,7 @@ describe('test app routes', () => {
         expect(object.apiVersion).toBe(fixtures.tokenReviewTemplate.apiVersion);
         expect(object.kind).toBe(fixtures.tokenReviewTemplate.kind);
         expect(object.status.authenticated).toBe(true);
-        expect(object.status.user).toEqual(fixtures.tokenPayload);
+        expect(object.status.user).toEqual(fixtures.tokenReviewResponse);
       })
       .end((error) => {
         done(error);
@@ -137,6 +124,53 @@ describe('test app routes', () => {
       .post('/token')
       .send({})
       .expect(400)
+      .end((error) => {
+        done(error);
+      });
+  });
+  test('GET /metrics: 200 OK (enabled basic auth)', (done) => {
+    request(app)
+      .post('/metrics')
+      .auth(config.prometheus.username, config.prometheus.password)
+      .expect(200)
+      .end((error) => {
+        done(error);
+      });
+  });
+  test('GET /metrics: 200 OK (with disabled basic auth)', (done) => {
+    process.env.PROMETHEUS_USERNAME = '';
+    process.env.PROMETHEUS_PASSWORD = '';
+    request(app)
+      .post('/metrics')
+      .expect(200)
+      .end((error) => {
+        delete process.env.PROMETHEUS_USERNAME;
+        delete process.env.PROMETHEUS_PASSWORD;
+        done(error);
+      });
+  });
+  test('GET /metrics: 401 Unauthorized (No authorization header)', (done) => {
+    request(app)
+      .post('/metrics')
+      .expect(401)
+      .end((error) => {
+        done(error);
+      });
+  });
+  test('GET /metrics: 401 Unauthorized (Wrong username/password)', (done) => {
+    request(app)
+      .post('/metrics')
+      .auth('john.doe', 'secret')
+      .expect(401)
+      .end((error) => {
+        done(error);
+      });
+  });
+  test('GET /metrics: 401 Unauthorized (Wrong username/password)', (done) => {
+    request(app)
+      .post('/metrics')
+      .auth('john.doe', 'secret')
+      .expect(401)
       .end((error) => {
         done(error);
       });