+++ test/unit/api/userAuthentication.test.js
@@ -2,7 +2,7 @@ import winston from 'winston';
 import {getResponseMock, getRequestMock} from '../mock';
 import jwt from 'jsonwebtoken';
 import UserAuthentication from '../../../src/api/userAuthentication';
-import {Authenticator, Mapping} from '../../../src/ldap';
+import {Authenticator} from '../../../src/ldap';
 jest.mock('../../../src/ldap/authenticator');
 
 const fixtures = {
@@ -11,22 +11,6 @@ const fixtures = {
   username: 'john.doe',
   password: 'secret',
   authHeader: 'Basic am9obi5kb2U6c2VjcmV0',
-  groups: [
-    'test',
-  ],
-  extraFields: {
-    uidNumber: 1,
-    gidNumber: 10,
-  },
-  mapping: new Mapping(
-    'uid',
-    'uid',
-    'memberOf',
-    [
-      'uidNumber',
-      'gidNumber',
-    ]
-  ),
 };
 
 let authenticator = new Authenticator();
@@ -34,20 +18,10 @@ const userAuthentication = new UserAuthentication(
   authenticator,
   fixtures.lifetime,
   fixtures.key,
-  fixtures.mapping,
   winston);
 
 beforeEach(() => {
   authenticator.authenticated = true;
-  authenticator.attributes = {
-    uid: fixtures.username,
-    memberOf: fixtures.groups.map((group) => {
-      return 'cn=' + group + ',dc=example,dc=com';
-    }),
-    uidNumber: fixtures.extraFields.uidNumber,
-    gidNumber: fixtures.extraFields.gidNumber,
-  };
-  authenticator.getAttributesShouldThrowError = false;
 });
 
 describe('UserAuthentication.run()', () => {
@@ -64,8 +38,6 @@ describe('UserAuthentication.run()', () => {
       let token = jwt.decode(responseMock.send.mock.calls[0][0], {complete: true});
       expect(token.header.typ).toBe('JWT');
       expect(token.payload.username).toBe(fixtures.username);
-      expect(token.payload.uid).toBe(fixtures.username);
-      expect(token.payload.groups).toEqual(fixtures.groups);
     });
   });
 
@@ -123,15 +95,20 @@ describe('UserAuthentication.run()', () => {
       .toEqual(400);
   });
 
-  test('Sends 500 on internal server error (e.g. ldap error)', () => {
-    authenticator.getAttributesShouldThrowError = true;
+  test('Sends 500 on invalid key', () => {
     const requestMock = getRequestMock(
       '',
       {'Authorization': fixtures.authHeader}
     );
     const responseMock = getResponseMock();
 
-    return userAuthentication.run(requestMock, responseMock).then(() => {
+    const failingUserAuthentication = new UserAuthentication(
+      authenticator,
+      fixtures.lifetime,
+      undefined,
+      winston);
+
+    failingUserAuthentication.run(requestMock, responseMock).then(() => {
       expect(responseMock.sendStatus).toHaveBeenCalled();
       expect(responseMock.sendStatus.mock.calls[0][0])
         .toEqual(500);
@@ -144,16 +121,12 @@ describe('UserAuthentication.getToken()', () => {
     let now = Math.floor(new Date() / 1000);
 
     expect.hasAssertions();
-    return userAuthentication.getToken(fixtures.username).then((result) => {
-          let token = jwt.decode(result, {complete: true});
-          expect(token.header.typ).toBe('JWT');
-          expect(token.payload.username).toBe(fixtures.username);
-          expect(token.payload.uid).toBe(fixtures.username);
-          expect(token.payload.groups).toEqual(fixtures.groups);
-          expect(token.payload.extra).toEqual(fixtures.extraFields);
-          expect(token.payload.exp / 60).toBeCloseTo((now + fixtures.lifetime) / 60);
-          expect(jwt.verify(result, fixtures.key));
-    });
+    let result = userAuthentication.getToken(fixtures.username);
+    let token = jwt.decode(result, {complete: true});
+    expect(token.header.typ).toBe('JWT');
+    expect(token.payload.username).toBe(fixtures.username);
+    expect(token.payload.exp / 60).toBeCloseTo((now + fixtures.lifetime) / 60);
+    expect(jwt.verify(result, fixtures.key));
   });
 });